refactor: cache import
