+++ internal/http/handler/blob.go
@@ -6,9 +6,9 @@ import (
 	"path/filepath"
 	"time"
 
-	"github.com/blobtrtl3/trtl3/internal/cache"
 	"github.com/blobtrtl3/trtl3/internal/domain"
 	"github.com/blobtrtl3/trtl3/internal/engine"
+	"github.com/blobtrtl3/trtl3/internal/infra/cache"
 	"github.com/blobtrtl3/trtl3/internal/queue"
 	"github.com/blobtrtl3/trtl3/shared"
 	"github.com/gin-gonic/gin"
